third_party/blink/web_tests/fast/events/mouse-click-events-expected.txt

This tests what mouse events we send.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Left Mouse Button
PASS eventLog is "mousedown(0) mouseup(0) click(0) mousedown(0) mouseup(0) click(0) dblclick(0) "
Middle Mouse Button
PASS eventLog is "mousedown(1) mouseup(1) auxclick(1) mousedown(1) mouseup(1) auxclick(1) "
Right Mouse Button
PASS eventLog is "mousedown(2) mouseup(2) auxclick(2) mousedown(2) mouseup(2) auxclick(2) "
PASS successfullyParsed is true

TEST COMPLETE
